feat: refactor the rules engine
