+++ src/handlers/transfers/handler.js
@@ -57,10 +57,15 @@ const decodePayload = Util.StreamingProtocol.decodePayload
 const Comparators = require('@mojaloop/central-services-shared').Util.Comparators
 const ErrorHandler = require('@mojaloop/central-services-error-handling')
 const { Ilp } = require('@mojaloop/sdk-standard-components')
+const SettlementModelRulesEngine = require('../../models/rules/settlement-model-rules-engine')
+const EnumCached = require('../../lib/enumCached')
+const SettlementModelCached = require('../../models/settlement/settlementModelCached')
 
 const consumerCommit = true
 const fromSwitch = true
 const toDestination = true
+
+const engine = new SettlementModelRulesEngine()
 
 /**
  * @function TransferPrepareHandler
@@ -110,6 +115,22 @@ const prepare = async (error, messages) => {
       const transactionObject = (new Ilp({ secret: null })).getTransactionObject(payload.ilpPacket)
       message.value.content.transaction = transactionObject
     }
+
+    // Select settlement model here
+    const allSettlementModels = await SettlementModelCached.getAll()
+    let settlementModels = allSettlementModels.filter(model => model.currencyId === message.value.content.payload.amount.currency)
+    if (settlementModels.length === 0) {
+      settlementModels = allSettlementModels.filter(model => model.currencyId === null) // Default settlement model
+      if (settlementModels.length === 0) {
+        throw ErrorHandler.Factory.createFSPIOPError(ErrorHandler.Enums.FSPIOPErrorCodes.GENERIC_SETTLEMENT_ERROR, 'Unable to find a matching or default, Settlement Model')
+      }
+    }
+    let settlementModel = settlementModels.find(sm => sm.ledgerAccountTypeId === Enum.Accounts.LedgerAccountType.POSITION)
+    if (Config.ENABLED_SETTLEMENT_MODEL_RULES_ENGINE && message.value.content.transaction) {
+      const ledgerAccountTypes = await EnumCached.getEnums('ledgerAccountType')
+      settlementModel = await engine.obtainSettlementModelFrom(message.value.content.transaction, settlementModels, ledgerAccountTypes)
+    }
+    message.value.content.settlementModel = settlementModel
 
     const headers = message.value.content.headers
     const action = message.value.metadata.event.action
@@ -187,7 +208,7 @@ const prepare = async (error, messages) => {
         Logger.isInfoEnabled && Logger.info(Util.breadcrumb(location, { path: 'validationPassed' }))
         try {
           Logger.isInfoEnabled && Logger.info(Util.breadcrumb(location, 'saveTransfer'))
-          await TransferService.prepare(payload)
+          await TransferService.prepare(message)
         } catch (err) {
           Logger.isInfoEnabled && Logger.info(Util.breadcrumb(location, `callbackErrorInternal1--${actionLetter}6`))
           Logger.isErrorEnabled && Logger.error(`${Util.breadcrumb(location)}::${err.message}`)
@@ -212,7 +233,7 @@ const prepare = async (error, messages) => {
         Logger.isErrorEnabled && Logger.error(Util.breadcrumb(location, { path: 'validationFailed' }))
         try {
           Logger.isInfoEnabled && Logger.info(Util.breadcrumb(location, 'saveInvalidRequest'))
-          await TransferService.prepare(payload, reasons.toString(), false)
+          await TransferService.prepare(message, reasons.toString(), false)
         } catch (err) {
           Logger.isInfoEnabled && Logger.info(Util.breadcrumb(location, `callbackErrorInternal2--${actionLetter}8`))
           Logger.isErrorEnabled && Logger.error(`${Util.breadcrumb(location)}::${err.message}`)
